fixed resolution problems in the ctrl-cut core

+++ src/ctrl-cut/config/DocumentSettings.hpp
@@ -38,7 +38,7 @@ public:
   const static Key<string>  DATA_DIR;
   const static Key<string>  FILENAME;
 
-  const static Key<uint16_t> RESOLUTION;
+  const static Key<double> RESOLUTION;
   const static Key<Distance> WIDTH;
   const static Key<Distance> HEIGHT;
   const static Key<bool> AUTO_FOCUS;
